feat: Add max age init option
In order to use the middleware now, a new instance needs to be created.
Should be more flexible in the future compared to just passing in
options directly to the middleware function. In the future, we need to
check that ServeHTTP is not called without next, or better yet, either
call next or proxy depending on the configuration.