fix: Do not cache when max-age too low
Do not set default max age in cache itself, making it possible to
opt-out of default max age altogether in CLI. (Set default max age to
zero).

Do not cache when max age is zero, but also not when max age less than
update timeout. The latter causes an infinite update loop, since the
response is always expiring in the loop. This check only applies when
updates are not turned off.