fix: Vary match and Content-Encoding match when identity
The earlier implementation reported not matching header fields when no
Vary and Content-Encoding fields were specified. This because the
initial value of the "match" was false. Now that value gets set to false
only when there actually are Vary and Content-Encoding headers, in which
case the default is to return false.

Also fixed situation where the response has identity (= no) encoding, in
which case a match is now reported. Although the client might disallow
identity encoding, this almost never happens, and is good enough for
now.