feat: Rewrite origin URL in text responses
Sometimes URLs in HTML and CSS are absolute URLs, which means demo'ing
acache is difficult (resources get loaded from the origin directly).
This commit adds the possibility to rewrite those URLs to an arbitrary
URL - usually the acache proxy server URL.

This is done with a CLI flag. If this flag is set, all upstream requests
are made with identity encoding (to disable compression). Then, when
saving the response, the URLs are rewritten (find and replace). This
means the initial request will still have the origin URLs, but this is
just for demo purposes, so should be fine.