fix: Crash when trying to update key with malformed URI
Fixes rare bug whereby a malformed URI that ended up as a key in the
cache caused the request creation to return an error (which was ignored)
and a request pointer of nil. This caused a crash when the update loop
encountered said key and URI.

Now all request creation codepaths have error checking to ensure the
request is properly created. Also, now updates that result in an error
will result in the corresponding cache entry to be purged.