fix: Short keys falsely matched with prefix
Always add extra tab character at the end of the URI, so that a request
for `/` (with key `GET:/`) does not match the saved response for `/foo`,
which we previously got by searching with the prefix. Now the keys are
`GET:/\t` and `GET:/foo\t` respectively, so the former is not a prefix
of the latter.