fix: Get origin-prefixed keys only when updating all
Previously the updateAll method would try to update all keys found in
the cache - even those not related to the AlwaysCache instance (i.e.
origin) in question. This commit changes the keys callback of the cache
provider to always require a prefix, as well as a prefix property on the
CacheKeyer.