feat: Stream proxied responses via ResponseWriter
Use httputils.Reverse proxy and the ResponseWriter tee to make proxied
requests streaming and hopefully simpler and more reliable.

Now also saving request and response time to DB directly for the sake of
simplicity. (We need expiration in the DB directly anyway for the
updates, so might as well save these other dates as well.)

Trying to create a cleaner public API for the caching RFC package.